### 2018-10-04 #257 @hettling
Submission: nbaR
The package is a full R client to the Netherlands Biodiversity API (NBA, see http://docs.biodiversitydata.nl/en/latest/) giving researchers access to several large Natural History and Botany collections in the Netherlands. Additionally, some convenience functions feature integration with other packages.

```
Package: nbaR
Title: R Package Client for the Netherlands Biodiversity API
Version: 0.0.0
Authors@R: c(person("Hannes", "Hettling", email = "hannes.hettling@naturalis.nl", role = c("aut", "cre")),
             person("Rutger", "Vos", role="aut"))
Maintainer: Hannes Hettling <hannes.hettling@naturalis.nl>
Description: Access to the digitised Natural History collection at the Naturalis Biodiversity Center.
    This is the official client to the Netherlands Biodiversity API
    (NBA, <http://api.biodiversitydata.nl>) for the R programming language.
    More information on the NBA can be found at <http://docs.biodiversitydata.nl>.
    tic,
    remotes,
    pkgdown
    httr,
    R6,
    ape
```

https://github.com/naturalis/nbaR

The package fits into the *data retrieval* category as it provides API access to the databases
at the Naturalis Biodiversity Center, Leiden, the Netherlands. Since R is a very important language in the ecology/evolutionary biology research community, this package gives low-threshold access to our data.

Target audience are ecologists, evolutionary biologists, taxonomists, and researchers interested in museum collections.

No.

**Note**: There is not yet a full manuscript since this requires collaboration with other departments in our institution. The time-allocation for writing the paper is not controlled by our group, we hence decided not to wait until a full paper is prepared with our collaborators.

`R CMD check` succeeds without warnings or notes.

Dom Bennett @dombennett Alexander Zizka @azizka
