### 2018-10-12 #257 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @hettling    !

Here's the output from [goodpractice](https://github.com/MangoTheCat/goodpractice/). If you haven't used `goodpractice` it's an R package that checks a number of things with another package - most of which we agree with and want authors to follow. You don't need to address these now, it's info for reviewers to use to get started. There's very little in the report below, which is a good thing!

```r
── GP nbaR ─────────
It is good practice to

    R/ApiClient.r:49:NA
    R/ApiClient.r:53:NA
    R/ApiClient.r:64:NA
    R/ApiClient.r:65:NA
    R/ApiClient.r:66:NA
    ... and 679 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    tests/testthat/test-GatheringEvent.r:92:1
    tests/testthat/test-MultiMediaGatheringEvent.r:92:1
    tests/testthat/test-MultiMediaObject.r:106:1
    tests/testthat/test-MultiMediaObject.r:108:1
    tests/testthat/test-Specimen.r:103:1
```

Seeking reviewers now 🕐

---

Reviewers:

- xxx
- xxx
