### 2018-10-12 #257 @sckott
@hettling are you going to be okay with transferring to ropensci github org, or do you need to retain in naturalis?
