### 2018-10-16 #257 @hettling
@sckott I'm ok to transfer it to ropensci!
