### 2018-11-08 #257 @DomBennett
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Sorry for the delay of my review, the package is bigger than I initially thought!

I think the package is very effective at what it sets out to do: it provides all the necessary tools for interacting with the NBA API, it is richly documented with details of all the classes and functions, and, in virtually all of my interactions with the package it functioned without any problems or errors. I therefore would very much suggest it be admitted to ROpenSci.

I did, however, struggle to review this package due to large parts of it being auto generated with swagger. Although I think it is really important to use a tool like swagger -- indeed it is recommended by ROpenSci -- to capture an updating service like an API, it does make the reviewer’s life a little difficult because I don’t know whether problem code is originating from the developer or the codegen engine.

To give an example, I found the process of querying the API a little annoying: generate a client object, generate query condition objects followed by a query specification object and, then, the results object. I understand that this R6-based process is created by the codegen, not the authors themselves. I can also see that the process is very powerful as by following so closely the design of the NBA API, it allows users to create any kind of query they want.

(Quick sidenote, as far as my Googling takes me, I think this may be the first swagger R package to be submitted to ROpenSci  -- with the possible exception of fishbaseapi?! I wasn’t able to find any similar R packages on which to emulate my own review.)

I fear though that your average bio/eco R user will find the R6 objects a little alien. I had to follow the articles and documentation very closely before I was able to make my own queries that were distinct from those provided as examples. My main requests, as a reviewer, would therefore be:

**Improve the documentation: more examples and better explanations**

There is a lot of documentation already, and it was really critical for me to be able to start making more of my own queries. I think though I would have liked a little more organisation in the “getting started” article and more real-world examples. I think the shark vignette is great because it provides a real world use for the package, additional articles need not be so long though. In particular I think it would be great if the “Getting started” article were split up (e.g. into real basics and then independent articles for each data domain). I think a visual schematic showing the basic class structures would really help as well. Finally, I was interested in the operators (equals, not equals, contains….) but found these were not sufficiently explained. More examples would help.

**Function wrappers?**

I have a thought that for the most basic of queries you could simply have wrapper functions, e.g.  `specimen_search` or `taxon_search`, which would take just names/ids and return R data structures that the majority of R users would understand. These wrappers would likely lead to much less coding on the part of the end-user and, although less flexible, they should make the package much more accessible.

E.g. this below wrapper let's a user run of the examples from "getting started".

```r
specimen_search <- function(value, field) {
  qcs <- vector(mode = 'list', length = length(value))
  for (i in seq_along(value)) {
    qcs[[i]] <- QueryCondition$new(field = field[[i]], operator = 'EQUALS',
                                   value = value[[i]])
  }
  qs <- QuerySpec$new(conditions = qcs)
  sc <- SpecimenClient$new()
  res <- sc$query(querySpec = qs, size = 100)
  lapply(X = res$content$resultSet, FUN = function(x) {
    x$item$toList()
  })
}

res <- specimen_search(value = c('female', 'species', "Equidae"),
                       field = c('sex', "identifications.taxonRank",
                                 "identifications.defaultClassification.family"))
```

Below I highlight some additional specific issues relating to automated checks and my general play-arounds.

#### Test results

When I first downloaded the repo, I got no errors when testing the package. But since 8 Nov 2018, I started getting the error and warnings below. This may just be due to the NBA API itself, but is it possible for the package be able to note the difference between API failure and internal failure?

```
test-multimediaClient-metadata.R:36: warning: Settings work
Status code:500
Internal Server Error
Exception: Invalid NBA setting: 50000
Exception type: java.lang.IllegalArgumentException
Full stack trace stored in response object
test-specimenClient-find.R:57: warning: Test error handling in find functions
404 (NOT FOUND)
No Specimen exists with ID XXX
test-specimenClient-misc.R:90: failure: getDistinctValuesPerGroup works
res$content has length 3, not length 2.
test-specimenClient-query.R:149: warning: Errors and warnings work
Status code:500
Internal Server Error
Exception: Field associatedMultiMediaUris.accessUri cannot be queried: field is not indexed
Exception type: nl.naturalis.nba.api.InvalidConditionException
Full stack trace stored in response object
test-taxonClient-find.R:35: warning: Test error handling in find functions
404 (NOT FOUND)
No Specimen exists with ID XXX
test-taxonClient-query.R:80: warning: Errors and warnings work
Status code:500
Internal Server Error
Exception: Invalid element "somefield" in path "somefield"
Exception type: nl.naturalis.nba.api.InvalidConditionException
Full stack trace stored in response object
══ Results ════════════════════════════════════════

OK:       1420
```

#### Check results

```
Warning: @noRd [/Users/djb208/Coding/nbaR/R/Utils.r#27]: has no parameters
Warning: @noRd [/Users/djb208/Coding/nbaR/R/Utils.r#205]: has no parameters

* checking installed package size ... NOTE
    R     1.4Mb
    doc   4.3Mb

* checking top-level files ... NOTE
Non-standard file/directory found at top level:
  ‘tic.R’
```

If the authors wish to submit to CRAN they will likely need to ensure that the package size is less than 5MB. The excessive size of the package is likely due to the images in the vignettes. They could prevent this by halting the building of vignettes in the R package itself or removing the images. I prefer the former option as I feel they could, in fact do with more use-cases (as explained above and below), and they can always refer users to the online documentation in the locally available R help files. If they choose the former, they should refer to this guide: https://pkgdown.r-lib.org/articles/pkgdown.html#articles

#### Goodpractice results

By removing the vignettes from the R package itself, you may also fix these issues goodpractice raised:

```
  ✖ write unit tests for all functions, and all package code in general.
    93% of code lines are covered by test cases.

    R/Feature.r:86:NA
    R/Feature.r:111:NA
    R/Feature.r:112:NA
    R/GeoArea.r:122:NA
    R/GeoArea.r:173:NA

  ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF version.
    This typically indicates Rd problems.
  ✖ fix this R CMD check ERROR: Re-running with no redirection of
    stdout/stderr. Hmm ... looks like a package You may want to clean up by 'rm -Rf
    /var/folders/ps/g89999v12490dmp0jnsfmykm0043m3/T//Rtmpn15kYK/Rd2pdf11fbd7d461aa9'
  ✖ fix this R CMD check NOTE: installed size is 6.3Mb sub-directories of
    1Mb or more: R 1.8Mb doc 4.3Mb
```

#### Spell check

I highlight a few spelling errors (en.US):

```
Baeysian                          sharks.Rmd:22
conventience                      sharks.Rmd:160
digitised                         nbaR.Rd:12
                                         description:1
donw                              tomato.Rmd:156
errenous                          sharks.Rmd:247
exaclty                           tomato.Rmd:62
expleore                          sharks.Rmd:147
objetcs                           sharks.Rmd:171
respose                           tomato.Rmd:47
retreive                          geo_age.Rd:19
                                  tomato.Rmd:28
whch                              tomato.Rmd:62
worning                           sharks.Rmd:236
```

#### Pkgdown issues

```
Warning: In '_pkgdown.yml', topic must be a valid R expression.
Problem topic: `\`Element\``
Warning: In '_pkgdown.yml', topic must be a valid R expression.
Problem topic: `\`GeoAreaClient\``
Warning: In '_pkgdown.yml', topic must be a valid R expression.
Problem topic: `\`Element\``
Warning: In '_pkgdown.yml', topic must be a valid R expression.
Problem topic: `\`GeoAreaClient\``
Warning: Topics missing from index: chronos_calib, geo_age
```

These above errors can be fixed by updating the pkgdown yaml. (Unless you are using tic to auto-generate the yaml?)

Additionally, it might be nice to organise the functions by utility to a user (e.g. `SpecimentClient`, `TaxonClient` should take precedence over `Agent`). In packages I have written in the past I have organised the functions/classes by public and private categories.

Also, Changelog leads to a 404.

#### Help files

In one of the vignettes, it was suggested I “Have a look at ?Response to see what this object contains.” -- yet the help file had little to no information. I think the level of information found in this help file is not too different from that of others.

I imagine that much of the documentation will have been generated with the code generator but i think it may be necessary to provide more information particularly for the more commonly used functions/classes. In particular, there are very few examples provided for the majority of the help files. I think it is required that all public functions/classes come with examples for any ROpenSci package.

With roxygen, examples can be stored in an examples/ folder and called in using roxygen syntax.

#### Vignettes/articles

**Too big**
I think it may be a good idea to split the main “get started” vignette into multiple separate vignettes. The initial size of the document can be off-putting and all of the available resources may not be useful for everyone who may wish to use nbaR. I would suggest breaking down by data type: specimen, taxon … etc. and having an article for each domain.

Additionally, the size of all the articles could be massively reduced by preventing much of the unnecessary printing to console of warnings, download status, long method lists by either clever behind-the-scenes coding or using suppressMessages/Warnings.

**sc?**
At this point in the text ….
`Note: The function getFieldInfo on a certain field lists which operators are allowed for that field (e.g. sc$get_field_info()$content$unitID$allowedOperators).`
I got confused as to what `sc` was. (Also is it getFieldInfo or get_field_info?)  I think this section could be improved. Perhaps a better example showing how a user might look up all the available query options for the specimen service.

```r
allowed_operators <- sc$get_field_info()$content
names(allowed_operators)
```

Or, perhaps a wrapper function?

**Schematic?**

It might be nice in the getting started vignette to provide a schematic of the query steps to highlight the commonalities (e.g. $new, $resultset … etc.)

```
# specify data service
client <- DATASERVICE$new()
# create query conditions
condition1 <- QueryCondition$new()
condition2 <- QueryCondition$new()
...
# specify query
query <- QuerySpec$new(conditions(list(condition1, condition2 ...))
# search
res <- res$query(querySpec = query)
```

**Map of tomatoes**

There is code, some of which was commented out, placed at the bottom of the En Tibi vignette without any accompanying text. Running this code I was able to generate a map of collected S. lycopersicum specimens. Is there a reason for it to be excluded? I am a little confused with the query for ‘gatheringEvent.siteCoordinates’ -- they are set to ‘NOT_EQUALS’ yet return only specimens which contain values in these slots.

**More information on operators**

It would be great if the different operators could be demonstrated. For example, I wanted to download the taxonomic records for a given list of genera. Initially I was trying to compile multiple query conditions using ‘OR’. But I since discovered that it could be easily done using the ‘IN’ operator with value as a vector of genera.

I noticed there are lots of operators available… would it be too much to ask for more examples that involve other operators other than just EQUALS?

#### Errors/issues

**Named list**

Named lists in query conditions do not work?

```r
library(nbaR)
sc <- SpecimenClient$new()
field <- 'identifications.defaultClassification.genus'
qc <- QueryCondition$new(field = field, operator = 'EQUALS', value = 'Solanum')
# success
qs <- QuerySpec$new(conditions = list(qc))
res <- sc$query(querySpec = qs)
res$content$totalSize
# fail
qc_list <- list('solanum' = qc)
qs <- QuerySpec$new(conditions = qc_list)
res <- sc$query(querySpec = qs)
res$content$totalSize
```

**Size**

I may have missed something here, but for me the `size` argument didn’t work -- the results were always stuck at 10.

```r
library(nbaR)
sc <- SpecimenClient$new()
qc <- QueryCondition$new(field = 'identifications.defaultClassification.genus',
                         operator = 'EQUALS', value = 'Solanum')
qc2 <- QueryCondition$new(field = 'identifications.defaultClassification.specificEpithet',
                          operator = 'EQUALS', value = 'lycopersicum')
qs <- QuerySpec$new(conditions = list(qc, qc2))
res <- sc$query(querySpec = qs, size = 1000)
res$content$totalSize == length(res$content$resultSet)
```

#### Final notes

- In the en tibi vignette, the first paragraph is repeated.
- Shark vignette: `get_age` is not a function. Should be `geo_age`?
- Shark vignette: broken images?
- Might be better to save the shark phylogeny as an R object (see `save` and use xz compression for CRAN ). That way the file can be compressed and loading is easier for a user: `data(shark_phylogeney)`. See: http://r-pkgs.had.co.nz/data.html
- It is good practice to spell out logicals: T and F should be TRUE and FALSE. (T and F are not protected)
- Keep 80 column width in vignette examples please.

#### My session info

```
Session info ---------------------------------------------------------------------------------
 system   x86_64, darwin15.6.0
 ui       RStudio (1.1.442)
 collate  en_GB.UTF-8
 tz       Europe/Stockholm

Packages -------------------------------------------------------------------------------------
 package   * version date       source
 ape         5.1     2018-04-04 CRAN (R 3.5.0)
 base      * 3.5.1   2018-07-05 local
 compiler    3.5.1   2018-07-05 local
 curl        3.2     2018-03-28 CRAN (R 3.5.0)
 datasets  * 3.5.1   2018-07-05 local
 devtools    1.13.6  2018-06-27 CRAN (R 3.5.0)
 digest      0.6.18  2018-10-10 cran (@0.6.18)
 graphics  * 3.5.1   2018-07-05 local
 grDevices * 3.5.1   2018-07-05 local
 grid        3.5.1   2018-07-05 local
 httr        1.3.1   2017-08-20 CRAN (R 3.5.0)
 jsonlite    1.5     2017-06-01 CRAN (R 3.5.0)
 lattice     0.20-35 2017-03-25 CRAN (R 3.5.1)
 memoise     1.1.0   2017-04-21 CRAN (R 3.5.0)
 methods   * 3.5.1   2018-07-05 local
 nbaR      * 0.0.0   2018-11-03 local
 nlme        3.1-137 2018-04-07 CRAN (R 3.5.1)
 parallel    3.5.1   2018-07-05 local
 R6          2.3.0   2018-10-04 cran (@2.3.0)
 Rcpp        0.12.18 2018-07-23 CRAN (R 3.5.0)
 stats     * 3.5.1   2018-07-05 local
 tools       3.5.1   2018-07-05 local
 utils     * 3.5.1   2018-07-05 local
 withr       2.1.2   2018-03-15 CRAN (R 3.5.0)
 yaml        2.2.0   2018-07-25 CRAN (R 3.5.0)
```

_In sum, I have no doubt that this package will do everything I’d expect as a portal to NBA. I would just like a bit more documentation, clearer explanation and more examples.
Thanks for inviting me to review!_
