### 2018-11-12 #257 @hettling
Thanks @DomBennett and @mbjoseph for thorough reviews of the package!
