### 2018-11-13 #257 @sckott
I don't think we have a deadline. If it will be a long time, we usually assign a `holding` tag so we know we don't have to worry about a submission until the person has time again. I'd guess every few weeks if we don't hear anything we ping whoever we're waiting on.
