### 2018-11-19 naturalis/nbaR#20 @hettling
Check results
Warning: @noRd [/Users/djb208/Coding/nbaR/R/Utils.r#27]: has no parameters
Warning: @noRd [/Users/djb208/Coding/nbaR/R/Utils.r#205]: has no parameters

* checking installed package size ... NOTE
  installed size is  5.9Mb
  sub-directories of 1Mb or more:
    R     1.4Mb
    doc   4.3Mb

* checking top-level files ... NOTE
Non-standard file/directory found at top level:
  ‘tic.R’
If the authors wish to submit to CRAN they will likely need to ensure that the package size is less than 5MB. The excessive size of the package is likely due to the images in the vignettes. They could prevent this by halting the building of vignettes in the R package itself or removing the images. I prefer the former option as I feel they could, in fact do with more use-cases (as explained above and below), and they can always refer users to the online documentation in the locally available R help files. If they choose the former, they should refer to this guide: https://pkgdown.r-lib.org/articles/pkgdown.html#articles
