### 2018-11-19 naturalis/nbaR#31 @hettling
 Final notes
- [Done] In the en tibi vignette, the first paragraph is repeated.
- [Done] Shark vignette: get_age is not a function. Should be geo_age?
- [Done] Shark vignette: broken images?
- [Done] Might be better to save the shark phylogeny as an R object (see save and use xz compression for CRAN ). That way the file can be compressed and loading is easier for a user: data(shark_phylogeney). See: http://r-pkgs.had.co.nz/data.html
- [Done] It is good practice to spell out logicals: T and F should be TRUE and FALSE. (T and F are not protected)
- [Done] Keep 80 column width in vignette examples please.
