### 2018-11-19 naturalis/nbaR#35 @hettling
Typos in the swagger README 
(https://github.com/naturalis/nbaR/blob/master/other/swagger/README.md)
"NAB" should be "NBA"
"overwiev"
"contatining"
