### 2018-11-19 naturalis/nbaR#36 @hettling
Specimen class docs
Further fields include (among others) information about Finding place, identification, and multimedia content.

It seems like "Finding" should not be capitalized (but maybe this is a swagger issue?)
