### 2018-11-19 naturalis/nbaR#37 @hettling
Character length in example code (vignettes)
Some of these code chunks have very long lines (over 80 char). It would be great to have shorter lines so that users don't need to scroll right.
