### 2018-11-19 naturalis/nbaR#38 @hettling
Query services example

https://github.com/naturalis/nbaR/blob/master/vignettes/nbaR.Rmd#L232-L246

This example is great - and may be a good substitute for the current first example of an empty query.
