### 2018-11-19 naturalis/nbaR#41 @hettling
find_by_ids
https://github.com/naturalis/nbaR/blob/master/vignettes/nbaR.Rmd#L333-L339

This method could be more idiomatic if it took a vector of ids rather than one string of ids separated by a comma.
