### 2018-11-19 naturalis/nbaR#44 @hettling
DwCA download services
It would be worth suppressing the output of the download to improve readability.
