### 2018-11-19 naturalis/nbaR#46 @hettling
Typos shark vignette
"Baeysian"
"Maximum-Likelihood" can be lowercase: "maximum likelihood" or "maximum-likelihood"
"For many specimen," should be "specimens"
"levels; If ...", capitalization of "If" is not necessary with the semicolon
"...there are sufficient specimen determined", should be plural "specimens"
