### 2018-11-19 naturalis/nbaR#47 @hettling
Shark vignette: Explaining the size argument for new users
https://github.com/naturalis/nbaR/blob/master/vignettes/sharks.Rmd#L89-L92

The comment here suggests that size is important - it could be nice to add some text explaining why this is necessary in this case, and how users can determine whether it is necessary in their own applications.
