### 2018-11-19 naturalis/nbaR#48 @hettling
Explaining warnings
https://github.com/naturalis/nbaR/blob/master/vignettes/sharks.Rmd#L190

This command raised two warnings:

> times <- geo_age(unique(c(as.character(data$youngChronoName), as.character(data$oldChronoName))))
Warning messages:
1: In FUN(X[[i]], ...) :
  Could not retreive values for geo unit "Mioceen" from earthlifeconsortium.org
2: In FUN(X[[i]], ...) :
  Could not retreive values for geo unit "Unspecified age" from earthlifeconsortium.org
It would be nice to have some text here for users to explain what these warnings are, and whether they matter.
