### 2018-11-20 naturalis/nbaR#19 @hettling
Now doing a 'ping' if API is live in test cases.
