### 2018-11-20 naturalis/nbaR#44 @hettling
This is now suppressed in the vignette output; if run in the console, httr puts a carriage return in the download progress, so there is no excessive output
