### 2019-01-03 naturalis/nbaR#48 @hettling
I think the warning messages do explain very well what happens: earthlifeconsortium.org cannot provide time values for these geological periods. I added the information that NA's are returned:

```
> geo_age(c('aa', 'bb', 'Miocene'))
          aa bb Miocene
early_age NA NA   23.03
late_age  NA NA    5.33
Warning messages:
1: In FUN(X[[i]], ...) :
  Could not retrieve time values for geo unit "aa" from earthlifeconsortium.org, returning NA
2: In FUN(X[[i]], ...) :
  Could not retrieve time values for geo unit "bb" from earthlifeconsortium.org, returning NA
>
```
The `In FUN(X[[i]])` we get because the function is vectorised.
