### 2019-01-04 naturalis/nbaR#29 @hettling
This works now, unit test with above example added.
