### 2019-01-29 naturalis/nbaR#37 @hettling
Done
