### 2019-01-29 naturalis/nbaR#31 @hettling
All points have been addressed.
