### 2019-01-29 naturalis/nbaR#18 @hettling
They should yet be documented in the vignettes!
