### 2019-02-28 #257 @hettling
@sckott Ok, reasonable. I would like to mention that there is progress in applying the reviewer's comments, documented in [this milestone](https://github.com/naturalis/nbaR/milestones). Remaining issues are mostly documentation related.
