### 2019-03-08 naturalis/nbaR#40 @hettling
In the wrapper and full client functions, this can be accomplished with
`list("sex"="female", "identifications.taxonRank"="species")`, which is I think
not so difficult compared to QueryConditions.
A more idiomatic syntax would be nice, but I think is more difficult to accomplish,
it would include parsing R formulas or expressions, I guess. Also, some error handling would
have to be implemented; for some fields, operators such as `<` or `>` are allowed, for others not.
I therefore will put a wontfix label.
