### 2019-04-11 naturalis/nbaR#20 @hettling
Runs without warnings or notes now, but will leave this open to re-run `R CMD check` before submission
