### 2019-04-11 naturalis/nbaR#26 @hettling
The *getting started* vignette now only contains examples about wrapper functions which are more straightforward. The *advanced queries* vignette has detailed examples, I think we thus do not need this schematic overview anymore.
