### 2019-04-16 naturalis/nbaR#34 @hettling
At least for the most important data types (Specimen, Taxon, GeoArea, MultiMediaObject) there are now good descriptions. The superfluous `Usage` sections are removed.
