### 2019-04-16 naturalis/nbaR#24 @hettling
Good descriptions have been added for the main datatypes: `Specimen`, `Taxon`, `GeoArea`, `Multimedia`.  Also the response and client classes are now properly documented. There is, still not a full description for each object model, since these are parsed out of code annotations from the NBA itsself. They would have to be added there by our developers.

However,  by looking at the fields, the user can pretty much see what the class is about. Furthermore, these classes are mere containers for data - there is not much functionality that needs to be described.
