### 2019-04-16 naturalis/nbaR#24 @hettling
Now also done for `QuerySpec` and `QueryCondition` objects.
