### 2019-04-16 naturalis/nbaR#17 @hettling
The main vignette is now split:

* Get Started (wrappers)
* Objects and advanced queries
* services summary
* sharks example
* tomato example
