### 2019-04-18 naturalis/nbaR#21 @hettling
A few 'long lines of code' and test coverage still missing, but the severe things are fixed. Therefore closing this now.
