### 2019-04-18 naturalis/nbaR#20 @hettling
Check now proceeds without errors, warnings or notes.
