### 2019-05-14 #257 @sckott
@DomBennett @mbjoseph any thoughts? if no response by the end of the week i'll assume you're happy with the changes
