### 2019-05-15 #257 @DomBennett
Hi @hettling

Looks good Hannes! That's an amazingly thorough response to every point raised in the reviews.

I re-ran `goodpractice` and got "NOTE: installed size is 5.9Mb sub-directories of 1Mb or more: R 2.0Mb doc 3.3Mb". This confuses me a little as the total repo size is less than 3.9 MB. Are the vigenettes particularly large upon installation? This will own be a real problem for you if you try and put the package on CRAN but it does also slow down the installation time. One way I've cut package sizes down in the past is to exclude vignettes from the build.
