### 2019-05-15 #257 @DomBennett
Also... I just noticed you have this kind of code at the beginning of a couple of your tests:

```r
wd <- getwd()
if (grepl("testthat", wd)) {
  data_dir <- file.path("data")
} else {
  ## for running test at package level
  data_dir <- file.path("tests", "testthat", "data")
}

tc <- TaxonClient$new(basePath = "http://api.biodiversitydata.nl/v2")
if (!tc$ping()) {
  skip("NBA not available, skipping test")
}
```

I recently learned about `testthat`'s setup and teardown functionalities. For shared variables like `data_dir` you could add a `setup-vars.R` script in your `testthat/` with the above code. This script would then be run at the beginning of each test script -- avoiding duplicate code.
