### 2019-05-15 #257 @sckott
thanks for your feedback @DomBennett

also for package size, if you have any datasets in thepkg, you can resave datasets with different compression options to see what is the smallest.
