### 2019-05-17 #257 @hettling
Hi @DomBennett ,

Strangely I do not get the note from goodpractice about package size. I run it as follows:

`gp = goodpractice::gp()`

and then if I look for the checks that contain 'size':

`goodpractice::results(gp)[c(164, 165, 211),]`

```
  check result
164            rcmdcheck_pdf_file_sizes   TRUE
165         rcmdcheck_pdf_file_sizes_gs   TRUE
211 rcmdcheck_reasonable_installed_size   TRUE`
```
If I check the size of the package in my library location, I get `4.0K`. During the revisions I excluded the shark and tomato examples from the vignettes, so now there are only three relatively small vignettes, without any figures. Am I missing something in calling goodpractice (maybe vignettes are not built etc.)?
