### 2019-05-21 ropensci/nbaR#58 @hettling
Ropenscience TODOs
* **DONE** i think you want to add paper.md and tic.R to your .Rbuildignore file
* **DONE** also maybe add vignettes/sharks.Rmd and vignettes/tomato.Rmd to your .Rbuildignore file as r cmd check complains about those files missing a vignette engine, which CRAN checks would also catch I think

* **DONE** Please transfer the package to ropensci- I've invited you to a team (you need to accept that invitation first, before transferring). You'll be made admin once you transfer
* **DONE** add rOpenSci footer to README
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
* **DONE** Change any needed links, such those for CI badges
* **DONE** Travis CI should update to the new location automatically - you may have to update other CI systems manually
* **DONE** If you use Appveyor, manage that under your own account
* **DONE** it's a good idea to add issue and PR template files to your .github folder, egs here https://github.com/ropensci/dotgithubfiles/ - and can be added with https://ropenscilabs.github.io/rodev/reference/use_ro_github.html
